# frontend/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Bitmoi service",
        "title": "Bitmoi API",
        "contact": {
            "name": "API Support",
            "email": "bitmoiigson@gmail.com"
        },
        "license": {
            "name": "CC BY-NC-SA 4.0",
            "url": "https://creativecommons.org/licenses/by-nc-sa/4.0/"
        },
        "version": "1.0"
    },
    "host": "api.bitmoi.co.kr",
    "basePath": "/",
    "paths": {
        "/bidToken": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "erc20"
                ],
                "summary": "광고 스팟에 MOI 토큰을 입찰합니다.",
                "parameters": [
                    {
                        "description": "입찰 금액과 광고 스팟",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BidTokenRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ScoreResponse"
                        }
                    }
                }
            }
        },
        "/competition": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chart"
                ],
                "summary": "경쟁모드에서 제공할 차트를 불러옵니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "제외할 USDT페어들을 쉼표로 구분하여 전달합니다.",
                        "name": "names",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OnePairChart"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "score"
                ],
                "summary": "경쟁모드에서 작성한 주문을 제출합니다.",
                "parameters": [
                    {
                        "description": "주문 정보",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ScoreRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ScoreResponse"
                        }
                    }
                }
            }
        },
        "/freeToken": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "erc20"
                ],
                "summary": "프리 MOI 토큰을 지급합니다.",
                "parameters": [
                    {
                        "description": "metamask 주소",
                        "name": "MetamaskAddressRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MetamaskAddressRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/highestBidder": {
            "get": {
                "tags": [
                    "erc20"
                ],
                "summary": "특정 광고 스팟에 가장 높은 가격을 제시한 입찰자를 반환합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "광고 스팟",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "최상위 입찰자",
                        "schema": {
                            "$ref": "#/definitions/api.HighestBidderResponse"
                        }
                    }
                }
            }
        },
        "/interval": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chart"
                ],
                "summary": "다른 시간단위의 차트를 불러옵니다. 연습, 경쟁 두 모드 모두 지원합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "identifier",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "competition",
                            "practice"
                        ],
                        "type": "string",
                        "name": "mode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "5m",
                            "15m",
                            "1h",
                            "4h",
                            "1d"
                        ],
                        "type": "string",
                        "name": "reqinterval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 10,
                        "minimum": 1,
                        "type": "integer",
                        "name": "stage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OnePairChart"
                        }
                    }
                }
            }
        },
        "/moreinfo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rank"
                ],
                "summary": "사용자가 랭크에 등재하며 기입한 추가 정보를 불러옵니다.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "score_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Score"
                            }
                        }
                    }
                }
            }
        },
        "/myscore/{page}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "score"
                ],
                "summary": "사용자의 경쟁모드 주문 채결 내역을 불러옵니다.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Score"
                            }
                        }
                    }
                }
            }
        },
...
        "/practice": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chart"
                ],
                "summary": "연습모드에서 제공할 차트를 불러옵니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "제외할 USDT페어들을 쉼표로 구분하여 전달합니다.",
                        "name": "names",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OnePairChart"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "score"
                ],
                "summary": "연습모드에서 작성한 주문을 제출합니다.",
                "parameters": [
                    {
                        "description": "주문 정보",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ScoreResponse"
                        }
                    }
                }
            }
        },
        "/rank": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rank"
                ],
                "summary": "사용자를 랭크에 등재합니다.",
                "parameters": [
                    {
                        "description": "랭크 등재 요청에 대한 정보",
                        "name": "rankInsertRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RankInsertRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rank/{page}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rank"
                ],
                "summary": "랭크에 등재된 사용자들을 불러옵니다.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.RankingBoard"
                            }
                        }
                    }
                }
            }
        },
        "/reissueAccess": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "access token을 재발급 합니다.",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "ReissueAccessTokenRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReissueAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReissueAccessTokenResponse"
                        }
                    }
                }
            }
        },
        "/selectedBidder": {
            "get": {
                "tags": [
                    "erc20"
                ],
                "summary": "특정 광고 스팟의 입찰에 성공한 사용자를 불러옵니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "광고 스팟",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "사용자와 입찰금액",
                        "schema": {
                            "$ref": "#/definitions/api.HighestBidderResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자를 생성합니다.",
                "parameters": [
                    {
                        "description": "유저 생성 요청에 필요한 정보",
                        "name": "CreateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/address": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자의 Metamask 계좌를 업데이트 합니다.",
                "parameters": [
                    {
                        "description": "Metamask 주소",
                        "name": "LoginUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MetamaskAddressRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/checkId": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "ID 중복 여부를 확인합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "체크할 ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/checkNickname": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "닉네임 중복 여부를 확인합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "체크할 닉네임",
                        "name": "nickname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BidTokenRequest": {
            "type": "object",
            "required": [
                "amount",
                "location"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                },
                "location": {
                    "type": "string",
                    "enum": [
                        "practice",
                        "rank",
                        "freetoken"
                    ]
                }
            }
        },
...
        "api.OnePairChart": {
            "type": "object",
            "properties": {
                "btcratio": {
                    "type": "number"
                },
                "entry_price": {
                    "type": "number"
                },
                "entrytime": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "onechart": {
                    "$ref": "#/definitions/api.CandleData"
                }
            }
        },
...
        "api.ScoreResponse": {
            "type": "object",
            "properties": {
                "origin_chart": {
                    "$ref": "#/definitions/api.CandleData"
                },
                "result_chart": {
                    "$ref": "#/definitions/api.CandleData"
                },
                "score": {
                    "$ref": "#/definitions/api.OrderResult"
                }
            }
        },
...
    }
}